Reimplemented configuration API in Rust (#4)
This replaces the old configuration API implementation with a new one written in Rust. Unlike the old implementation, which was entirely Windows-specific, the new implementation is platform-agnostic, well-defined in terms of supported variables, and can easily be adapted to new configuration file formats.

In addition, this also changes the configuration file format from INI to TOML. TOML is similar to INI, but is formally specified and has several advantages over INI, such as nesting and additional value types. Other formats such as JSON or YAML could have been used instead, but TOML was chosen specifically because TOML documents resemble INI files, so users will likely find them familiar enough to work with.

The implementation itself is divided into two components: a library providing the actual implementation, and a transition layer that emulates the old API. The latter is intended solely for existing C++ code and will hopefully be phased out someday.